+++ Altinn/AT.Common.Altinn.Publish/Implementation/Clients/AltinnAuthenticationClient.cs
@@ -23,6 +23,7 @@ internal class AltinnAuthenticationClient : IAltinnAuthenticationClient
                 .Get($"exchange/{tokenProvider.ToString().ToLower()}")
                 .WithBearerToken(tokenProviderToken)
                 .WithAcceptHeader("text/plain", 1.0)
-                .ReceiveString() ?? throw new Exception("Failed to exchange token with Altinn");
+                .ReceiveString()
+            ?? throw new Exception("Failed to exchange token with Altinn");
     }
 }