+++ AspNetCore/AT.Common.AspNetCore.Publish/CHANGELOG.md
@@ -19,6 +19,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security <!-- in case of vulnerabilities. -->
 
+## 2.2.0
+
+### Changed
+
+- changed(deps): Applied minor and patch updates to dependencies
+
+## 2.1.3
+
+### Added
+
+- chore(aspnetcore): Use two distinct health endpoints for readiness and liveness as default (`healthz/ready` and `healthz/live`) instead of just one for all. Also, add a default `StartupHealthCheck` which can be configured to run multiple startup tasks like e.g. a database migration before it will return a healthy state.
+
+## 2.1.2
+
+### Fixed
+
+- fix(scalar): Ensure scalar is using https when requested with https. Http otherwise
+
+## 2.1.1
+
+### Fixed
+
+- fix(scalar): Downgrade scalar because v2.11.0 has a UI bug which leads to unexpected
+
 ## 2.1.0
 
 ### Changed