+++ Enhetsregisteret/AT.Common.Enhetsregisteret.Publish/CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security <!-- in case of vulnerabilities. -->
 
+## 1.1.0
+
+### Changed
+
+- changed(deps): Applied minor and patch updates to dependencies
+
 ## 1.0.1
 
 ### Changed